Refactoring code (#34)
Quiz Parser parses the html form of the quiz and validates it for script attacks.
Quiz checker to be called after quiz is attempted to compare answers and calculate result.
Storage handlers interact with the dropbox. Downloading and uploading files with relevant names.
@controller is the decorator which handles alerts. So some methods in the view do not have this decorator that is because they do not render any pages but only redirect. So they do not have alerts to be displayed so no decorator for them.
There is a controller folder. This will have all the controllers later.
Corresponding changes in test cases.